requirements.txt
Django==2.2.7
django-extensions==2.2.5
djangorestframework==3.10.3

# dj-database-url==0.5.0
psycopg2==2.8.4